Add a step to make sure operators are actually installed
The current task does not check the status of operators to make
sure they are successfully deployed. It further breaks the
openstack services deployment by giving unwanted errors.

This patch adds a step to make sure operators are successfully
installed. It will help to fail the job when there is an issue
with the meta operator installation.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>